Updates: 7th round New library: cli_prompt.php New library: is_float_equal.php New library: var_export_contains.php New test: cli_prompt.php New test: is_float_equal.php New test: measure_exec_time.php New test: print_file.php.phtml New test: queue_worker.php New test: sleep.js.phtml New test: var_export_contains.php New tool: get-composer.php New tool: matthiasmullie-minify.php Updated component: herring Updated component: middleware_form Updated HOWTO.md Updated library: addDesktopIcon.js Updated library: beautify.js Updated library: button.css Updated library: cache_container.php Updated library: checkDate.js Updated library: cron.php Updated library: enableTabOnTextarea.js Updated library: getCookie.js Updated library: getJson.js Updated library: ioc_container.php Updated library: list2tree.js Updated library: logger.php Updated library: pdo_cheat.php Updated library: pdo_connect.php Updated library: richTextEditor.js Updated library: school_algorithms.php Updated library: sec_bruteforce.php Updated library: sec_captcha.php Updated library: sec_login.php Updated library: sec_lv_encrypter.php Updated library: singleton.php Updated library: sleep.js Updated public/.htaccess Updated README.md Updated samples Updated test: addDesktopIcon.js.phtml Updated test: array_tree.php Updated test: beautify.js.phtml Updated test: blog_page_slicer.php Updated test: cache_container.php Updated test: checkDate.js.phtml Updated test: cron.php Updated test: curl_file_updown.php Updated test: directoryIterator_sort.php Updated test: enableTabOnTextarea.js.phtml Updated test: file_http_request.php Updated test: getallheaders.php Updated test: http_request_response.php Updated test: list2tree.js.phtml Updated test: logger.php Updated test: multipage.js.phtml Updated test: pdo_cheat.php Updated test: pdo_connect.php Updated test: pdo_crud_builder.php Updated test: richTextEditor.js.phtml Updated test: school_algorithms.php Updated test: sec_bruteforce.php Updated test: sec_captcha.php Updated test: sec_lv_encrypter.php Updated test: simpleblog_db.php Updated test: superclosure.php Updated test: time_converter.php Updated tool: cron.php Updated tool: logrotate.php Updated tool: opcache-preload-generator.php Updated tool: run-php-components-tests.php Updated tool: serve.php Updated tool: webdevsh.php Updated tool: websockets.php

+++ tests/logger.php
@@ -5,6 +5,21 @@
 	 * Note:
 	 *  looks for a library at ../lib
 	 *
+	 * Hint:
+	 *  you can setup database credentials by environment variables
+	 *  variables:
+	 *   TEST_DB_TYPE (pgsql, mysql, sqlite, overrides first argument)
+	 *   TEST_PGSQL_HOST (default: 127.0.0.1)
+	 *   TEST_PGSQL_PORT (default: 5432)
+	 *   TEST_PGSQL_DBNAME (default: php_toolkit_tests)
+	 *   TEST_PGSQL_USER (default: postgres)
+	 *   TEST_PGSQL_PASSWORD (default: postgres)
+	 *   TEST_MYSQL_HOST (default: [::1])
+	 *   TEST_MYSQL_PORT (default: 3306)
+	 *   TEST_MYSQL_DBNAME (default: php-toolkit-tests)
+	 *   TEST_MYSQL_USER (default: root)
+	 *   TEST_MYSQL_PASSWORD
+	 *
 	 * Warning:
 	 *  PDO extension is required
 	 *  pdo_sqlite extension is required
@@ -74,30 +89,74 @@
 				exit(1);
 			}
 
+		if(getenv('TEST_DB_TYPE') !== false)
+			$argv[1]=getenv('TEST_DB_TYPE');
 		if(isset($argv[1]))
 		{
-			switch($argv[1])
-			{
-				case 'pgsql':
-					$pdo_handler=new PDO('pgsql:'
-						.'host=127.0.0.1;'
-						.'port=5432;'
-						.'dbname=logger_test;'
-						.'user=postgres;'
-						.'password=postgres'
-					);
-				break;
-				case 'mysql':
-					$pdo_handler=new PDO('mysql:'
-						.'host=[::1];'
-						.'port=3306;'
-						.'dbname=logger-test',
-						'root',
-						''
-					);
+			$_db_type=$argv[1];
+			$_db_credentials=[
+				'pgsql'=>[
+					'host'=>'127.0.0.1',
+					'port'=>'5432',
+					'dbname'=>'php_toolkit_tests',
+					'user'=>'postgres',
+					'password'=>'postgres'
+				],
+				'mysql'=>[
+					'host'=>'[::1]',
+					'port'=>'3306',
+					'dbname'=>'php-toolkit-tests',
+					'user'=>'root',
+					'password'=>''
+				]
+			];
+			foreach(['pgsql', 'mysql'] as $database)
+				foreach(['host', 'port', 'dbname', 'user', 'password'] as $parameter)
+				{
+					$variable='TEST_'.strtoupper($database.'_'.$parameter);
+					$value=getenv($variable);
+
+					if($value !== false)
+					{
+						echo '  -> Using '.$variable.'="'.$value.'" as '.$database.' '.$parameter.PHP_EOL;
+						$_db_credentials[$database][$parameter]=$value;
+					}
+				}
+
+			try {
+				switch($_db_type)
+				{
+					case 'pgsql':
+						if(!extension_loaded('pdo_pgsql'))
+							throw new Exception('pdo_pgsql extension is not loaded');
+
+						$pdo_handler=new PDO('pgsql:'
+							.'host='.$_db_credentials[$_db_type]['host'].';'
+							.'port='.$_db_credentials[$_db_type]['port'].';'
+							.'dbname='.$_db_credentials[$_db_type]['dbname'].';'
+							.'user='.$_db_credentials[$_db_type]['user'].';'
+							.'password='.$_db_credentials[$_db_type]['password'].''
+						);
+					break;
+					case 'mysql':
+						if(!extension_loaded('pdo_mysql'))
+							throw new Exception('pdo_mysql extension is not loaded');
+
+						$pdo_handler=new PDO('mysql:'
+							.'host='.$_db_credentials[$_db_type]['host'].';'
+							.'port='.$_db_credentials[$_db_type]['port'].';'
+							.'dbname='.$_db_credentials[$_db_type]['dbname'],
+							$_db_credentials[$_db_type]['user'],
+							$_db_credentials[$_db_type]['password']
+						);
+				}
+			} catch(Throwable $error) {
+				echo ' Error: '.$error->getMessage().PHP_EOL;
+				exit(1);
 			}
 
-			$pdo_handler->exec('DROP TABLE log');
+			if(isset($pdo_handler))
+				$pdo_handler->exec('DROP TABLE log');
 		}
 		if(!isset($pdo_handler))
 			$pdo_handler=new PDO('sqlite:'.__DIR__.'/tmp/logger.sqlite3');
